SAE: Flush PMKSA if an assoc reject without timeout is received
Flush the PMKSA upon receiving association reject event without timeout
in the event data in SME-in-driver case to avoid trying to use the old
PMKSA entry in subsequent connection attempts. Do not flush PMKSA if
association reject is received with timeout as it is generated
internally from the driver without reaching the AP. This is similar to
the SME-in-wpa_supplicant case that was already addressed within
sme_event_assoc_reject().

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Git-commit: d109aa6cacf2c3f643de0c758a30b0daf936a67a
Git-repo: git://w1.fi/srv/git/hostap.git
CRs-fixed: 2308932
Change-Id: I7f76da6adeee7c49c4bce66091f592c778d4ba3d
[This addresses the fix done in commit e729793f09943038e7ee0a55598b212628470c50
 by having the current_ssid checked prior to the access]