eap_proxy: Use eloop context for re-initialization during modem ssr.
eap_proxy registers and initializes uim service(s) as part of eap_proxy
initialization in separate thread via eap_proxy_post_init as it might
take time. This helps to break the context from supplicant, which
could continue with initialization. While the above behavior works good
for the wpa_supplicant initialization, the same is followed during the
SSR and this causes the issue described below.

SSR would trigger eap_proxy_qmi_deinit followed by eap_proxy_post_init,
which runs in a separate pthread. First thing eap_proxy_qmi_deinit does
is to join the thread for eap_proxy_post_init to ensure the initialization
happens through that thread before de-initializing the same.

With the above behavior, if the SSR and wpa_supplicant's deinit occur in
a quick time and same order, the sequence would be a eap_proxy_qmi_deinit,
create a thread for eap_proxy_post_init and eap_proxy_deinit. On some
platforms, we have observed that the assignment of the thread_id as part
of the thread creation is not synchronous and thus both the deinit's end
up joining the same thread_id, which would result in unexpected
behavior(SIGABRT).

Thus, do not create create a thread for eap_proxy_post_init during SSR
(this would avoid the context from getting broken) and join only for
the first deinit.

Change-Id: Ie74fd9b9e347b85ab4beb8e1e27b84d7d40f5fe7
CRs-Fixed: 2325413