nl80211: Disable the registration for the RRM frame
wpa_supplicant's registration for the RRM frame conflicts with the
one done by the LOWI. Hence, disable the registration from wpa_supplicant
as this functionaity of wpa_supplicant is not used by our architecture.

The ideal way to disable this is to control through the ctrl parameter
from the framework , which requires HIDL enhancements , which is a
long poll. As an interim, we are disabling this by default, since the
intention anyways is to disable this registration from wpa_supplicant.

Change-Id: I197d40d1875b120f17347eaf5b2135a4bd7a322e
CRs-Fixed: 2342472