SAE: Reuse previously generated PWE on a retry with the same STA
Do not start SAE authentication from scratch if a STA starts a new
attempt for the same group while we still have previously generated PWE
available. Instead, use the previously generated PWE as-is and skip
anti-clogging token exchange since the heavy processing is already
completed. This saves unnecessary processing on the AP side in case the
STA failed to complete authentication on the first attempt (e.g., due to
heavy SAE load on the AP causing a timeout) and makes it more likely for
a valid STA to be able to complete SAE authentication during a DoS
attack.

Git-repo : git://w1.fi/srv/git/hostap.git
Git-commit: f2143615810cc73e6369ff181e615536188a702d
Change-Id: If982e22d24bd898187136d3dd2e8a63214b25036
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
CRs-fixed: 2414695