SAE: Enable only group 19 by default in AP mode
Change the AP mode default for SAE to enable only the group 19 instead
of enabling all ECC groups that are supported by the used crypto library
and the SAE implementations. The main reason for this is to avoid
enabling groups that are not as strong as the mandatory-to-support group
19 (i.e., groups 25 and 26). In addition, this disables heavier groups
by default.

In addition, add a warning about MODP groups 1, 2, 5, 22, 23, and 24
based on "MUST NOT" or "SHOULD NOT" categorization in RFC 8247. All the
MODP groups were already disabled by default and would have needed
explicit configuration to be allowed.

Git-repo : git://w1.fi/srv/git/hostap.git
Git-commit: 941bad5ef4db4dc6436f2d321d40abcb936ba7dd
Change-Id: Id3ac9528a51cac800f0f9ea29fdba6fe353d3cd3
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
CRs-fixed: 2414695