SAE: Add support for using the optional Password Identifier
This extends the SAE implementation in both infrastructure and mesh BSS
cases to allow an optional Password Identifier to be used. This uses the
mechanism added in P802.11REVmd/D1.0. The Password Identifier is
configured in a wpa_supplicant network profile as a new string parameter
sae_password_id. In hostapd configuration, the existing sae_password
parameter has been extended to allow the password identifier (and also a
peer MAC address) to be set. In addition, multiple sae_password entries
can now be provided to hostapd to allow multiple per-peer and
per-identifier passwords to be set.

Git-repo : git://w1.fi/srv/git/hostap.git
Git-commit: 9be19d0b9c4e4948e70fbfeb9076d30af9d0071f
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
CRs-fixed: 2414700

Change-Id: Id540ffe2a6345f0fe549121baca99b76cae23e2b